[JENKINS-37791] refactor sort logic to utility
