+++ blueocean-personalization/src/main/js/util/SortUtils.js
@@ -1,0 +1,62 @@
+/**
+ * Created by cmeyers on 9/14/16.
+ */
+// the order the cards should be displayed based on their result/state (aka 'status')
+const statusSortOrder = [
+    'UNKNOWN', 'FAILURE', 'ABORTED', 'NOT_BUILT',
+    'UNSTABLE', 'RUNNING', 'QUEUED', 'SUCCESS',
+];
+
+const extractStatus = (favorite) => {
+    const latestRun = favorite && favorite.item && favorite.item.latestRun || {};
+    return latestRun.result === 'UNKNOWN' ? latestRun.state : latestRun.result;
+};
+
+// sorts the cards based on 1. status 2. endTime, startTime or enQueueTime (descending)
+export const sortByStatusByRecent = (favoriteA, favoriteB) => {
+    const statusA = extractStatus(favoriteA);
+    const statusB = extractStatus(favoriteB);
+    const orderA = statusSortOrder.indexOf(statusA);
+    const orderB = statusSortOrder.indexOf(statusB);
+
+    if (orderA < orderB) {
+        return -1;
+    } else if (orderA > orderB) {
+        return 1;
+    }
+
+    const endTimeA = favoriteA && favoriteA.item && favoriteA.item.latestRun && favoriteA.item.latestRun.endTime;
+    const endTimeB = favoriteB && favoriteB.item && favoriteB.item.latestRun && favoriteB.item.latestRun.endTime;
+
+    if (endTimeA && endTimeB) {
+        const endCompare = endTimeA.localeCompare(endTimeB);
+
+        if (endCompare !== 0) {
+            return -endCompare;
+        }
+    }
+
+    const startTimeA = favoriteA && favoriteA.item && favoriteA.item.latestRun && favoriteA.item.latestRun.startTime;
+    const startTimeB = favoriteB && favoriteB.item && favoriteB.item.latestRun && favoriteB.item.latestRun.startTime;
+
+    if (startTimeA && startTimeB) {
+        const startCompare = startTimeA.localeCompare(startTimeB);
+
+        if (startCompare !== 0) {
+            return -startCompare;
+        }
+    }
+
+    const queuedTimeA = favoriteA && favoriteA.item && favoriteA.item.latestRun && favoriteA.item.latestRun.enQueueTime;
+    const queuedTimeB = favoriteB && favoriteB.item && favoriteB.item.latestRun && favoriteB.item.latestRun.enQueueTime;
+
+    if (queuedTimeA && queuedTimeB) {
+        const queueCompare = queuedTimeA.localeCompare(queuedTimeB);
+
+        if (queueCompare !== 0) {
+            return -queueCompare;
+        }
+    }
+
+    return 0;
+};
